Add MarkupTextBlock and update filter function
Added a new control, MarkupTextBlock, to allow dynamically specifying markup content in a TextBlock. Multiple files were added or modified to support this feature, including an update to the filtering function in LsWorkingDataService which now utilizes regular expressions for robust filtering of data. Fixed a text icon spelling error in the MainWindow view.