Implemented MaterialDesign UI and extended search functionality
The UI has been refactored to implement the Material Design toolkit in the LSLocalizeHelper project which includes changes in the appearance and style of the MainWindow. This enhanced the overall user experience by providing a modern and sleek interface. Furthermore, previously used Dark and Light themes were removed as part of the refactoring.

The search functionality was also extended to include previous translations, providing a more comprehensive search experience. The hits in the search results are now highlighted in red color for more visibility.

In the existing implementation, the copy to clipboard function from Current and Previous Texts was disabled to avoid confusing redundant actions, retaining only the Copy for Translated Texts. The display area for Current and Previous Text was changed from TextBlocks to MarkupTextBlocks to include highlighting of search results.

The enhancements also cater to the addition of respective texts to the Translations DataGrid, replacing the former MarkupText Columns with explicit Origin, Translated and Previous.

A new attribute "Previous" was added in the DataRowModel to hold the previous translation text. Corresponding changes were made in the LsWorkingDataService to support this additional attribute.

The unnecessary logic of setting origin texts in MainWindow.actions.cs and MainWindow.commands.cs was removed to improve the performance and maintain the simplicity of the code.