Update directory referencing and error handling in Bg3ModsService
Updated the way the parent directory was referenced in Bg3ModsService. `metaFile.Directory?.Parent?.Parent?.Parent` is now used for better accuracy. This prevents the file from being placed in the wrong directory. Refactored the error handling in `BuildPackage` method, to check for a scarce number of invalid conditions rather than a high number of valid conditions. Also changed log output from `Debug.WriteLine` to `Console.WriteLine` to align with other methods, and removed unused references to streamline the application. In the app's resource file, locale-specific dictionaries are now uncommented, allowing the selection of different languages.