Add NotEmptyValidationRule to check for null or empty strings
This new ValidationRule subclass, NotEmptyValidationRule, has been introduced to ensure that key fields won't be left empty during data entry by users. It checks if the value is not null or whitespace and will return a validation error otherwise. This significantly improves the robustness of data validation in our localization helper.