Add import new mod functionality and refactor commands
Added a new function `DoImportMod()` to allow for importing of new mods, along with multiple commands refactored to use ICommand. Also improved data binding in the MainWindow XAML. The refactor and new function are part of enhancing application feature set and applying better practice for the command handling in MVVM framework.