Update localization keys and values in en.xaml
This commit expands and modifies the localization keys and respective values in the locals.en.xaml file. The changes stems from the need to enhance the localization capabilities of the app by including broader control keys (such as "ButtonAbort", "ButtonCopy", etc). Also, the commit adds translation text for different parts of the application like import mod function, selections, settings and version info. These better articulate the needs of the application in a more user-friendly way and adheres strictly to internationalization standards.