Improve error handling and refactor localization
Refactored how errors are handled, thrown, and logged in the LsWorkingDataService. Removed unused properties in ImportDialog.xaml.cs. Added a new helper class for handling resources in resources.cs. Shifted the application to a resource-based approach for better localization and removed unused imports across several .cs files. This change is necessary for logging and debugging related activities and also to provide multiple languages support for the application. Future error messages will be more informative and localized as per the user preference.