Update Chinese Translation in LSLocalizeHelper
Renamed locals.zh.xaml to locals.zh-Hans.xaml and revised translation of certain terms to be more precise and easier to understand. The changes were made to improve the user interface of the program in Simplified Chinese version. This ensures the users clearly understand the given instructions and options while using the LSLocalizeHelper program.