Replace XmlLanguage with CultureInfo for language localization
The XmlLanguage method had limitations for the localization process. As such, we replaced it with CultureInfo which is more accurate and accepts more languages. This change reflects in the MainWindow.model.cs and MainWindow.events.cs files.

Additionally, in MainWindow.xaml, we improved the binding structure of the DataGrid to enhance readability and maintainability. In MainWindow.xaml.cs, we updated the SetLocals method to accommodate the CultureInfo change and also set the UI and thread culture accordingly.

This will help create a more efficient and accurate localization process.