"Improve XML data handling and UI interaction
Refactored the way XML data is loaded and saved in MainWindow.actions.cs to enhance readability and maintainability. Functions were added to instantiate a new XML document from dataRowModels and to save it, alongside with a reworked mechanism for copying to clipboard upon row change. To improve user experience and interaction with the UI, corresponding functions and events were tweaked in MainWindow.commands.cs, MainWindow.events.cs, and MainWindow.xaml files. A template for new XML files was also created (local.xml) and embedded into the project."