Add StatusBarModel and update status bar functionality
Added a new StatusBarModel for managing conditions of the status bar. Extracted the status properties from the MainWindowModel to StatusBarModel for better encapsulation and separation of concerns. Now, with this change,the status bar is updated with the execution of each operation to reflect the current state of the application. This increases user-friendliness, allowing users to see current counts, changes and loading status immediately. StatusBarModel also follows the INotifyPropertyChanged pattern to automatically update the UI as its properties change.