Add function to apply translated text and update UI layout
Function 'DoApplyTranslatedText' was added to make it possible to set the translated text for a particular UUID and to refresh the status bar accordingly. The 'ButtonApplyTranslated_OnClick' and 'ButtonPasteTranslated_OnClick' were updated to use this new function. The layout of the MainWindow was also updated to improve the grouping and arrangement of components for better UX. This change will aid in making it easier to apply translations and view application status.