Add support for .net 7.0 and update application resources
Added .net 7.0 to the list of target frameworks in 'LSLocalizeHelper.csproj'. Also, favicon.ico is now set as the application icon and 'icon2.jpg' is included in the project. 'Splash.jpg' was also removed from the resources. These changes aim to modernize the application's look and feel, while providing compatibility with the latest .NET version.