Enable Windows targeting in project files
Changed the order of target frameworks and enabled Windows targeting by adding `<EnableWindowsTargeting>true</EnableWindowsTargeting>` in project files for LSLocalizeHelper, LsHelperUnitTests, and LsLocalizeHelperLib. This will allow the projects to be specifically targeted for a Windows environment during the build process, thereby ensuring better compatibility and performance when the application is run on Windows platforms.