Refactor namespace and update test assertion
Adjusted the namespace for test classes to be more specific, moving them from 'LsHelperUnitTests' to 'LsHelperUnitTests.Tests'. This change was done to better align with project structure and convention. Also, the assertion in the 'ModsServiceTests' was modified from checking strictly for 2 items to checking for more than 1 item, providing better flexibility with potential future changes to the test dataset.