Update GitHub actions and project versions
Updated '.github/workflows/dotnet.yml' to use .NET 6.0.x instead of 7.0.x and checkout@v4 and setup-dotnet@v3 instead of v2 and v1 respectively. The build verbosity is set to normal.

Also, a new file 'Directory.Build.props' is added to centralize the version management for the solution. As a consequence, redundant version properties are removed from individual project files.

This change ensures better control and uniformity over the solution version, and the updated GitHub actions boost the performance by using the stable .NET version and the latest actions.