Update dotnet framework versions and build verbosity
The dotnet framework version matrix was simplifying to only include 'net6.0-windows'. This was done to focus on the recent .NET 6.0 version for the time being and enable efficient CI/CD pipeline. The dotnet build verbosity level was also changed from 'normal' to 'minimal' to make the output of builds clearer and easier to understand.