Add GitHub tag name display in workflows
A new step is added to both 'push_tag.yml' and 'push.yml' GitHub workflows which will print the GitHub tag name. This was done to enable easier identification and tracking of the actions linked to specific tags, especially during debugging and log review. This modification echoes the 'GITHUB_REF_NAME' and 'github.ref_name', along with the part of 'GITHUB_REF' after the last slash.