Update GitHub workflows' names and fix a syntax error
The names of the GitHub Actions workflows were changed to provide clearer descriptions of when they run. The names were previously 'LsLocalHelper.NET' and are now 'On Pull Requests', 'On Push Branches', and 'On Push Tags' respectively. Also, a syntax error in the 'push_tag.yml' workflow file was fixed, specifically in the 'target_file' variable’s template literal syntax which caused build failure.