Update workflows and add tag referencing in the build job
In this update, I've commented out the test job and removed the dependency of build job on the test job in the push_tag.yml. I've added a 'tag' command for clearer tag referencing on the build job. In detecting the presence of new releases, the updated method now checks if an existing release has the same tag before creating a new one. It will also automatically add release notes on creation.

In pull_request.yml, the Restore dependencies, Build, and Test commands were initially commented out but have been re-enabled again.