Remove unnecessary Bg3LocaHelper.sln.DotSettings
The Bg3LocaHelper.sln.DotSettings file was removed because it was not necessary for the project. This file contained configuration settings specifically for ReSharper which are not required to be shared for the project. Its removal will lessen repository clutter and ensure that ReSharper settings remain customizable locally for individual developers.