Update GitHub release creation command in push_tag.yml
The command for creating a GitHub release in the push_tag.yml workflow file was simplified. The `--notes-from-tag` option was removed because it was redundant as release notes are directly fetched from the RELEASE.MD file. Instead, added a new variable `nightly_flags` to be used as additional flags for release creation. This makes the workflow more flexible and control over release notes clearer.