Update GitHub Actions to use latest Windows version
The workflow configuration for the GitHub Actions has been updated. Previously it used a specific version of Windows (windows-2023), now it has been changed to use the latest available version of Windows (windows-latest) for the "test" and "print-token" jobs. This change ensures our CI/CD process is always running on the most updated Windows environment.