Simplify environment variable handling and output in GitHub action
This commit simplifies the handling of environment variables in the push_tag GitHub workflow. The verbose printing of different forms of each variable has been reduced, printing the necessary ones only. This not only makes the code easier to understand and maintain but also minimizes potential points of failure or confusion.