Add sorting functionality to the 'TranslationGrid' in LsLocalizeHelper
In LsLocalizeHelper's MainWindow, we have added sorting functionality to the 'TranslationGrid'. This ensures the data in the grid can be sorted in any desired order. A new method 'RestoreOrder' was added to call 'sortingHelper.SyncToGrid()' at necessary points. This improvement enhances the UI's flexibility and improves the user experience by facilitating easier data searches and analysis.