Refactor LsTextBox control and MainWindow view for better readability
Changes made in LsTextBox.xaml and MainWindow.xaml files to enhance readability and maintainability. The properties in LsTextBox were rearranged in alphabetical order and the ScrollViewer properties were moved from MainWindow to LsTextBox for better cohesion and encapsulation. Furthermore, unnecessary spacing was eliminated and the CaretStoryBoard was re-organized for code clarity. This change does not affect the functionalities.