"Refactor for ReactiveUI integration and INPC simplification
Updates were made to Model classes to utilize ReactiveUI for change notifications. This refactor significantly simplifies the code and reduces boilerplate. Additionally, removed ViewModelBase class, as it's no longer needed with ReactiveUI's ReactiveObject as a base for models. The ReactiveUI.Fody package and necessary wiring were added to the csproj, enabling Fody's compile-time weaving for automatic implementation of INPC."