Update package versions and refactor translation functions
Updated several package versions in LsHelperUnitTests.csproj and LsLocalizeHelperLib.csproj for the latest features and security patches.

Also refactored translation functions in MainWindow.events.cs to reduce code duplication. Instead of having separate, almost identical functions for translating with Microsoft and MyMemory, a new function DoTranslate() was introduced that takes a TranslateType parameter to determine which API to use.

Additionally, made modifications in MainWindow.xaml for user interface enhancements including addition of progress bar and various design revisions.

Lastly, made changes in StatusBarModel.cs to implement a Reactive Property for the NotModified flag, enhancing the responsiveness of the UI.