"Add translation aborting functionality
Two string keys have been added to each localization file to support the new feature of aborting translations. In the MainWindow.actions.cs, code was added to handle the case when a translation process is aborted. An Abort button was also added in the UI in the MainWindow.events.cs file to have abortion functionality. In terms of the models, changes were made to MainWindow.model.cs including refactoring some variable assignments and addition of AbortPending flag. Scattered minor refactoring also occurred throughout the files to enhance code readability."